// src/utils/LE/globals/enums.ts
export enum EnTipoDato {
  noVariable = 0,
  uint8 = 1,
  int8 = 2,
  uint16 = 3,
  int16 = 4,
  uint32 = 5,
  int32 = 6,
  float = 7,
  tiempo = 8,

  pUint8 = 9,
  pInt8 = 10,
  pUint16 = 11,
  pInt16 = 12,
  pUint32 = 13,
  pInt32 = 14,

  pFloat = 15,
  pTiempo = 16,

  fecha = 26,

  string4 = 30,
  string32 = 34,

  // ? Este tipo de dato está en desuso: se utiliza TIPO_P_STRING_unicode16_simple que está más optimizado
  pString32 = 36,

  pStringUnicode16 = 44,
  stringUnicode16 = 45,

  concatenado = 46,
  pStringUnicode32 = 47,

  stringUnicode32 = 48,

  int64 = 202,
  pUint64 = 205,
  pFecha = 206,
}